--- src/main/kotlin/ru/spb/ivsamokhvalov/example/demo/camunda/service/Notify.kt
@@ -1,6 +1,0 @@
-package ru.spb.ivsamokhvalov.example.demo.camunda.service
-
-interface NotifyService {
-
-
-}
